almost got training working on jets
